chore: Remove mobile token restrictions
