+++ src/select-travel-token-screen/SelectTravelTokenScreenComponent.tsx
@@ -10,12 +10,7 @@ import {
 import {StyleSheet, Theme, useThemeContext} from '@atb/theme';
 import {ThemedTokenPhone, ThemedTokenTravelCard} from '@atb/theme/ThemedAssets';
 import {useFareContracts} from '@atb/ticketing';
-import {
-  dictionary,
-  getTextForLanguage,
-  TravelTokenTexts,
-  useTranslation,
-} from '@atb/translations';
+import {dictionary, TravelTokenTexts, useTranslation} from '@atb/translations';
 import {animateNextChange} from '@atb/utils/animation';
 import React, {useCallback, useEffect, useState} from 'react';
 import {ActivityIndicator, View} from 'react-native';
@@ -37,7 +32,7 @@ type Props = {onAfterSave: () => void};
 
 export const SelectTravelTokenScreenComponent = ({onAfterSave}: Props) => {
   const styles = useStyles();
-  const {t, language} = useTranslation();
+  const {t} = useTranslation();
   const {theme} = useThemeContext();
 
   const {disable_travelcard} = useRemoteConfigContext();
@@ -71,27 +66,17 @@ export const SelectTravelTokenScreenComponent = ({onAfterSave}: Props) => {
   );
 
   // Filter for unique travel rights config types
-  const availableFareContractsTypes = availableTravelRights.map(
-    (travelRight) => {
-      const preassignedFareProduct = findReferenceDataById(
-        preassignedFareProducts,
-        isOfFareProductRef(travelRight) ? travelRight.fareProductRef : '',
-      );
-
-      return (
-        preassignedFareProduct &&
-        fareProductTypeConfigs.find(
-          (c) => c.type === preassignedFareProduct.type,
-        )
-      );
-    },
-  );
-
-  const fareProductConfigWhichRequiresTokenOnMobile =
-    availableFareContractsTypes.find(
-      (fareProductTypeConfig) =>
-        fareProductTypeConfig?.configuration.requiresTokenOnMobile === true,
+  availableTravelRights.map((travelRight) => {
+    const preassignedFareProduct = findReferenceDataById(
+      preassignedFareProducts,
+      isOfFareProductRef(travelRight) ? travelRight.fareProductRef : '',
     );
+
+    return (
+      preassignedFareProduct &&
+      fareProductTypeConfigs.find((c) => c.type === preassignedFareProduct.type)
+    );
+  });
 
   useEffect(() => {
     // Whenever a user enters this screen, the onboarding is done.
@@ -121,14 +106,6 @@ export const SelectTravelTokenScreenComponent = ({onAfterSave}: Props) => {
 
   const travelCardToken = tokens?.find((t) => t.type === 'travel-card');
   const mobileTokens = tokens?.filter((t) => t.type === 'mobile');
-
-  // Shows an error message if switching to a t:card,
-  // but the current inspectable token is in the mobile AND
-  // requires mobile token
-  const requiresTokenOnMobile =
-    selectedType === 'travel-card' &&
-    inspectableToken?.type === 'mobile' &&
-    !!fareProductConfigWhichRequiresTokenOnMobile;
 
   return (
     <View style={styles.container}>
@@ -206,24 +183,6 @@ export const SelectTravelTokenScreenComponent = ({onAfterSave}: Props) => {
             isMarkdown={false}
           />
         )}
-        {requiresTokenOnMobile && (
-          <MessageInfoBox
-            type="error"
-            title={t(
-              TravelTokenTexts.toggleToken.notAllowedToUseTravelCardError.title,
-            )}
-            message={t(
-              TravelTokenTexts.toggleToken.notAllowedToUseTravelCardError.message(
-                getTextForLanguage(
-                  fareProductConfigWhichRequiresTokenOnMobile.name,
-                  language,
-                ) ?? '',
-              ),
-            )}
-            style={styles.errorMessageBox}
-            isMarkdown={false}
-          />
-        )}
         {selectedType === 'mobile' && mobileTokens?.length ? (
           <RadioGroupSection<Token>
             type="spacious"
@@ -254,16 +213,14 @@ export const SelectTravelTokenScreenComponent = ({onAfterSave}: Props) => {
         {toggleMutation.isLoading ? (
           <ActivityIndicator size="large" />
         ) : (
-          !requiresTokenOnMobile && (
-            <Button
-              expanded={true}
-              onPress={onSave}
-              text={t(TravelTokenTexts.toggleToken.saveButton)}
-              interactiveColor={theme.color.interactive[0]}
-              disabled={!selectedToken || (data?.toggleLimit ?? 0) < 1}
-              testID="confirmSelectionButton"
-            />
-          )
+          <Button
+            expanded={true}
+            onPress={onSave}
+            text={t(TravelTokenTexts.toggleToken.saveButton)}
+            interactiveColor={theme.color.interactive[0]}
+            disabled={!selectedToken || (data?.toggleLimit ?? 0) < 1}
+            testID="confirmSelectionButton"
+          />
         )}
       </ScrollView>
     </View>
