+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Ticketing_TicketTabNavStack/TicketTabNav_PurchaseTabScreen/TicketTabNav_PurchaseTabScreen.tsx
@@ -35,12 +35,9 @@ export const TicketTabNav_PurchaseTabScreen = ({navigation}: Props) => {
   const styles = useStyles();
   const analytics = useAnalyticsContext();
 
-  const {tokens, mobileTokenStatus} = useMobileTokenContext();
+  const {tokens} = useMobileTokenContext();
   const inspectableToken = tokens.find((t) => t.isInspectable);
   const hasInspectableMobileToken = inspectableToken?.type === 'mobile';
-  const hasMobileTokenError =
-    mobileTokenStatus === 'fallback' || mobileTokenStatus === 'error';
-
   if (must_upgrade_ticketing) return <UpgradeSplash />;
 
   const onProductSelect = (fareProductTypeConfig: FareProductTypeConfig) => {
@@ -55,7 +52,6 @@ export const TicketTabNav_PurchaseTabScreen = ({navigation}: Props) => {
     if (authenticationType !== 'phone') {
       if (
         fareProductTypeConfig.configuration.requiresLogin &&
-        fareProductTypeConfig.configuration.requiresTokenOnMobile &&
         !hasInspectableMobileToken
       ) {
         navigation.navigate('Root_LoginRequiredForFareProductScreen', {
@@ -70,24 +66,6 @@ export const TicketTabNav_PurchaseTabScreen = ({navigation}: Props) => {
         });
         return;
       }
-    } else if (
-      fareProductTypeConfig.configuration.requiresTokenOnMobile &&
-      !hasInspectableMobileToken &&
-      !hasMobileTokenError
-    ) {
-      navigation.navigate(
-        'Root_ActiveTokenOnPhoneRequiredForFareProductScreen',
-        {
-          nextScreen: {
-            screen: 'Root_PurchaseOverviewScreen',
-            params: {
-              selection,
-              mode: 'Ticket',
-            },
-          },
-        },
-      );
-      return;
     }
 
     navigation.navigate('Root_PurchaseOverviewScreen', {
